Reorganise record,task and contact packages
