+++ src/main/java/wallet/command/Command.java
@@ -1,18 +1,18 @@
 package wallet.command;
 
-import wallet.contact.Contact;
-import wallet.contact.ContactList;
-import wallet.record.Expense;
-import wallet.record.ExpenseList;
-import wallet.record.ExpenseParser;
-import wallet.record.RecordList;
+import wallet.model.contact.Contact;
+import wallet.model.contact.ContactList;
+import wallet.model.record.Expense;
+import wallet.model.record.ExpenseList;
+import wallet.model.record.ExpenseParser;
+import wallet.model.record.RecordList;
 import wallet.storage.Storage;
-import wallet.task.Deadline;
-import wallet.task.Event;
-import wallet.task.ScheduleList;
-import wallet.task.Task;
-import wallet.task.TaskList;
-import wallet.task.Tentative;
+import wallet.model.task.Deadline;
+import wallet.model.task.Event;
+import wallet.model.task.ScheduleList;
+import wallet.model.task.Task;
+import wallet.model.task.TaskList;
+import wallet.model.task.Tentative;
 
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
