+++ src/main/java/wallet/model/record/ExpenseList.java
@@ -1,4 +1,4 @@
-package wallet.record;
+package wallet.model.record;
 
 import java.util.ArrayList;
