+++ src/main/java/wallet/model/record/Notes.java
@@ -1,0 +1,4 @@
+package wallet.model.record;
+
+public class Notes {
+}
