+++ src/main/java/wallet/model/task/ScheduleList.java
@@ -1,4 +1,4 @@
-package wallet.task;
+package wallet.model.task;
 
 import java.util.ArrayList;
