+++ src/main/java/wallet/model/task/Tentative.java
@@ -1,9 +1,8 @@
-package wallet.task;
+package wallet.model.task;
 
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Date;
-import java.util.List;
 
 public class Tentative extends Task {
