--- src/main/java/wallet/record/Notes.java
@@ -1,4 +1,0 @@
-package wallet.record;
-
-public class Notes {
-}
