+++ src/test/java/wallet/model/task/TaskListTest.java
@@ -1,4 +1,4 @@
-package wallet.task;
+package wallet.model.task;
 
 import org.junit.jupiter.api.Test;
